refactored team component to account for new info
