+++ client/components/Team.jsx
@@ -4,12 +4,13 @@ import PropTypes from 'prop-types';
 
 function Team({columnNames, deletePlayer, savedPlayers}) {
   let keys = -1;
-  const elArr = [];
+
+  const columnHeadersArr = [];
   // create an array of column headers for the table
-  for (const key of columnNames) {
-    elArr.push(<th className="table-head" key={key}>{key}</th>)
+  for (const key in savedPlayers[0]) {
+    if(key === '_id' || key == '__v' ) continue;
+    columnHeadersArr.push(<th className="table-head" key={key}>{key}</th>)
   }
-  console.log('saved players: ', savedPlayers);
   // create rows for the table and fill out columns for each player
   const rows = [];
   if (savedPlayers[0] !== undefined) {
@@ -17,11 +18,11 @@ function Team({columnNames, deletePlayer, savedPlayers}) {
       const data = [];
       data.push(<td key={keys--}><button onClick={() => {
         let name =savedPlayers[i].name;
-        console.log( name)
         deletePlayer(name);
       }}>delete</button></td>)
       // inner for loop fills out all of the columns for one player/row
       for (const key in savedPlayers[i]) {
+        if(key === '_id' || key == '__v' ) continue;
         data.push(<td className="team-column" key={keys--}>{savedPlayers[i][key]}</td>)
       }
       rows.push(<tr className="team-data" key={keys--} data-testid='team-row'>{data}</tr>);
@@ -35,7 +36,7 @@ function Team({columnNames, deletePlayer, savedPlayers}) {
           <thead className="team-head">
             <tr>
               <th>action</th>
-              {elArr}    
+              {columnHeadersArr}    
             </tr>
           </thead>
           <tbody>
